[REFACTOR] rename search to music search
