+++ app/src/main/java/com/bluepig/alarm/ui/music/select/MediaSelectBottomSheetDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.bluepig.alarm.ui.search.select
+package com.bluepig.alarm.ui.music.select
 
 import android.os.Bundle
 import android.view.View
