+++ app/src/main/res/layout/fragment_music_search.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.search.MusicSearchFragment">
+    tools:context=".ui.music.MusicSearchFragment">
 
     <androidx.appcompat.widget.AppCompatEditText
         android:id="@+id/et_search"
